--- app-libs/node-api-types/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "ita-node-api-types"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-
-# scs
-substrate-api-client = { default-features = false, git = "https://github.com/scs/substrate-api-client", branch = "polkadot-v0.9.26" }
-
-# substrate
-sp-core = { optional = true, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.26" }
-
-
-[features]
-default = ["std"]
-std = [
-    "substrate-api-client/std",
-    "substrate-api-client/ws-client",
-    "sp-core"
-]
