+++ core-primitives/node-api/Cargo.toml
@@ -6,37 +6,13 @@ edition = "2018"
 
 [dependencies]
 # crates.io
-codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false, features = ["derive"] }
-thiserror = { version = "1.0", optional = true }
-
-# sgx dependencies
-sgx_types = { branch = "master", git = "https://github.com/apache/teaclave-sgx-sdk.git" }
-sgx_tstd = { branch = "master", git = "https://github.com/apache/teaclave-sgx-sdk.git", optional = true }
-
-# sgx-compatible crates
-thiserror_sgx = { package = "thiserror", git = "https://github.com/mesalock-linux/thiserror-sgx", tag = "sgx_1.1.3", optional = true }
-
-# substrate
-sp-core = { git = "https://github.com/paritytech/substrate.git", default-features = false, branch = "polkadot-v0.9.26" }
-sp-finality-grandpa = { git = "https://github.com/paritytech/substrate.git", default-features = false, branch = "polkadot-v0.9.26" }
-sp-runtime = { git = "https://github.com/paritytech/substrate.git", default-features = false, branch = "polkadot-v0.9.26" }
 
 # scs
 substrate-api-client = { default-features = false, git = "https://github.com/scs/substrate-api-client", branch = "polkadot-v0.9.26" }
 
-# integritee
-itp-types = { default-features = false, git = "https://github.com/integritee-network/pallets.git", branch = "master" }
-
 [features]
 default = ["std"]
 std = [
-    "itp-types/std",
     "substrate-api-client/std",
-    "thiserror",
+    "substrate-api-client/ws-client",
 ]
-sgx = [
-    "sgx_tstd",
-    "thiserror_sgx",
-]
-# used for unit testing only!
-mocks = []
