+++ core-primitives/node-api/src/lib.rs
@@ -15,59 +15,39 @@
 
 */
 
-//! Some substrate-api-client extension traits.
+//! All kind of things to talk to a node.
 
-#![cfg_attr(not(feature = "std"), no_std)]
-#![cfg_attr(test, feature(assert_matches))]
+//! Contains type definitions to talk to the node.
+///
+/// For example, you need to update this if you have a signed extensions in your node that
+/// is different from the integritee-node, or if you use the `pallet_asset_tx_payment`.
+use substrate_api_client::{
+	PlainTip, PlainTipExtrinsicParams, PlainTipExtrinsicParamsBuilder, SubstrateDefaultSignedExtra,
+	UncheckedExtrinsicV4,
+};
 
-#[cfg(all(feature = "std", feature = "sgx"))]
-compile_error!("feature \"std\" and feature \"sgx\" cannot be enabled at the same time");
+/// Configuration for the ExtrinsicParams.
+///
+/// Valid for the default integritee node
+pub type ParentchainExtrinsicParams = PlainTipExtrinsicParams;
+pub type ParentchainExtrinsicParamsBuilder = PlainTipExtrinsicParamsBuilder;
 
-#[cfg(all(not(feature = "std"), feature = "sgx"))]
-extern crate sgx_tstd as std;
+// Pay in asset fees.
+//
+// This needs to be used if the node uses the `pallet_asset_tx_payment`.
+//pub type ParentchainExtrinsicParams = AssetTipExtrinsicParams;
+//pub type ParentchainExtrinsicParamsBuilder = AssetTipExtrinsicParamsBuilder;
 
-// re-export module to properly feature gate sgx and regular std environment
-#[cfg(all(not(feature = "std"), feature = "sgx"))]
-pub mod sgx_reexport_prelude {
-	pub use thiserror_sgx as thiserror;
+pub type ParentchainUncheckedExtrinsic<Call> =
+	UncheckedExtrinsicV4<Call, SubstrateDefaultSignedExtra<PlainTip>>;
+
+#[cfg(feature = "std")]
+pub use api::*;
+
+#[cfg(feature = "std")]
+mod api {
+	use super::ParentchainExtrinsicParams;
+	use substrate_api_client::{rpc::WsRpcClient, Api};
+
+	pub type ParentchainApi = Api<sp_core::sr25519::Pair, WsRpcClient, ParentchainExtrinsicParams>;
 }
-
-#[cfg(all(not(feature = "sgx"), feature = "std"))]
-pub use substrate_api_client::{rpc::WsRpcClient, Api, ApiClientError};
-
-#[cfg(all(not(feature = "sgx"), feature = "std"))]
-pub mod account;
-
-#[cfg(all(not(feature = "sgx"), feature = "std"))]
-pub mod chain;
-
-#[cfg(all(not(feature = "sgx"), feature = "std"))]
-pub mod node_api_factory;
-
-#[cfg(all(not(feature = "sgx"), feature = "std"))]
-pub mod pallet_teerex;
-
-#[cfg(all(feature = "mocks", feature = "std"))]
-pub mod pallet_teerex_api_mock;
-
-pub mod error;
-pub mod metadata;
-
-#[cfg(all(not(feature = "sgx"), feature = "std"))]
-pub use account::*;
-#[cfg(all(not(feature = "sgx"), feature = "std"))]
-pub use chain::*;
-#[cfg(all(not(feature = "sgx"), feature = "std"))]
-pub use pallet_teerex::*;
-
-#[cfg(all(not(feature = "sgx"), feature = "std"))]
-use itp_types::ParentchainExtrinsicParams;
-
-#[cfg(all(not(feature = "sgx"), feature = "std"))]
-use sp_core::sr25519;
-
-#[cfg(all(not(feature = "sgx"), feature = "std"))]
-pub type ApiResult<T> = Result<T, ApiClientError>;
-
-#[cfg(all(not(feature = "sgx"), feature = "std"))]
-pub type ParentchainApi = Api<sr25519::Pair, WsRpcClient, ParentchainExtrinsicParams>;
